requirements.txt
# app
uvicorn==0.15.0
fastapi==0.70.0
pydantic==1.8.2
#db
pymongo==3.12.1

#dev
python-dotenv
pytest
pytest-dependency
pytest-asyncio
httpx
asgi-lifespan
mongomock